re-arranged folder
added deprecated folder to include old documents
replaced the old documents with new ones keeping the same names
