+++ aws/docker/worker/deprecated/requirements.txt
@@ -1,0 +1,6 @@
+requests==2.22.0
+matplotlib>=2.2.4
+cmocean==2.0
+netCDF4>=1.4.2
+netcdftime>=1.0.0a2
+boto3==1.13.18
