+++ aws/docker/worker/requirements.txt
@@ -1,6 +1,5 @@
-requests==2.22.0
-matplotlib>=2.2.4
+boto3==1.9.66
 cmocean==2.0
-netCDF4>=1.4.2
-netcdftime>=1.0.0a2
-boto3==1.13.18
+erddapy==0.5.3 
+matplotlib==3.1.3
+numpy==1.18.1
